// src/api-client/models/EventType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Type of event in the support system
 * @export
 */
export const EventType = {
  Email: "EVENT_EMAIL",
  SettingChange: "EVENT_SETTING_CHANGE",
  HubspotTicket: "EVENT_HUBSPOT_TICKET",
  Note: "EVENT_NOTE",
  Job: "EVENT_JOB",
  Skedulo: "EVENT_SKEDULO",
  ZendeskTicket: "EVENT_ZENDESK_TICKET",
} as const;
export type EventType = (typeof EventType)[keyof typeof EventType];

export function EventTypeFromJSON(json: any): EventType {
  return EventTypeFromJSONTyped(json, false);
}

export function EventTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): EventType {
  return json as EventType;
}

export function EventTypeToJSON(value?: EventType | null): any {
  return value as any;
}